chore: moved updateDetailsSectionSlice
