+++ src/widgets/BhandaraCheckin/SectionUpdateDetails/updateDetailsCheckin/checkinAbhyasi.ts
@@ -1,13 +1,11 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 import { RootState, ThunkApiConfig } from "../../store";
 import { snackbarSlice } from "../../../../components/Snackbar/snackbarSlice";
-import {
-  bhandaraCheckinSlice,
-  updateDetailsSectionSlice,
-} from "../../store/slices";
+import { bhandaraCheckinSlice } from "../../store/slices";
 import { postAttendance } from "../../store/api-async-thunks";
 import { UserDetails } from "../../types";
 import { RefinedCityStateCountryLocation } from "../../../../components/LocationTextField/locations";
+import { updateDetailsSectionSlice } from "../updateDetailsSectionSlice";
 
 const getEmailValue = (userDetails: UserDetails): { email?: string } => {
   if (userDetails.email.value?.match(/\*/)) {
